requirements.txt
pytest
allure-pytest
selene>=2.0.0rc4
pydantic_settings
Appium-Python-Client